requirements.txt
psycopg2
aiogram
aioschedule
peewee
peewee-migrations
python-dateutil